move remaining files to contrib to fix import layering errors
