+++ src/vs/workbench/api/browser/mainThreadNotebookEditors.ts
@@ -4,9 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 // --- Start Positron ---
-import { POSITRON_NOTEBOOK_EDITOR_ID } from '../../contrib/positronNotebook/common/positronNotebookCommon.js';
+import { POSITRON_NOTEBOOK_EDITOR_ID, usingPositronNotebooks } from '../../contrib/positronNotebook/common/positronNotebookCommon.js';
 import { checkPositronNotebookEnabled } from '../../contrib/positronNotebook/browser/positronNotebookExperimentalConfig.js';
-import { usingPositronNotebooks } from '../../services/positronNotebook/common/positronNotebookUtils.js';
 // --- End Positron ---
 import { DisposableStore, dispose } from '../../../base/common/lifecycle.js';
 import { equals } from '../../../base/common/objects.js';
