+++ src/vs/workbench/contrib/positronNotebook/browser/ContextKeysManager.ts
