--- src/vs/workbench/services/positronNotebook/common/positronNotebookUtils.ts
@@ -1,16 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2025 Posit Software, PBC. All rights reserved.
- *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { IConfigurationService } from '../../../../platform/configuration/common/configuration.js';
-
-/**
- * Check if Positron notebooks are configured as the default editor for .ipynb files
- * @param configurationService Configuration service
- * @returns true if Positron notebooks are the default editor, false otherwise
- */
-export function usingPositronNotebooks(configurationService: IConfigurationService): boolean {
-	const editorAssociations = configurationService.getValue<Record<string, string>>('workbench.editorAssociations') || {};
-	return editorAssociations['*.ipynb'] === 'workbench.editor.positronNotebook';
-}
